Added class Motor
