--- Work2/src/Motor.java
@@ -1,15 +1,0 @@
-public class Motor {
-    public int volume;
-
-    Motor(int volume) {
-        this.volume = volume;
-    }
-
-    @Override
-    public String toString() {
-        return "Motor" +
-                " volume = " + volume +
-                " cubic centimeters" +
-                ",";
-    }
-}
